Parser now preserves precision in calculations.

+++ src/com/ccl/utils/MathUtils.java
@@ -82,7 +82,7 @@ public class MathUtils
 		return Short.toString(temp);
 	}
 	
-	public static int getPercentageValue(float percentage, int min, int max)
+	public static int getPercentageValue(double percentage, int min, int max)
 	{
 		return (int) ((max - min) * percentage + min);
 	}
